requirements.txt
requests>=2.31
feedparser>=6.0
pyzotero>=1.5
notion-client>=2.0
wordcloud>=1.9
matplotlib>=3.8
pyyaml>=6.0
pytest>=8.0
trafilatura>=2.0
sentence-transformers>=3.0
peft>=0.10
numpy>=1.24
pymupdf>=1.24
